small fixes and testing
- test authentication
- fix issues with redux-persist
- fix backend cors issues

+++ frontend/lib/storage.ts
@@ -1,23 +1,36 @@
-import { configureStore } from "@reduxjs/toolkit";
-import { persistStore, persistReducer } from "redux-persist";
+import { Platform } from "react-native";
 import { MMKV } from "react-native-mmkv";
+import { configureStore } from "@reduxjs/toolkit";
+import {
+  persistStore, persistReducer, Storage,
+  FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER
+} from "redux-persist";
 
 import { rootReducer } from "./state";
 
+const isWeb = Platform.OS === "web";
+const isBrowser = typeof window !== 'undefined' && typeof window.localStorage !== 'undefined';
+
 const storage = new MMKV();
 
-const storageBackend = {
+const storageBackend: Storage = {
   setItem: async (key: string, value: string): Promise<boolean> => {
+    if (isWeb && !isBrowser) return Promise.resolve(false);
+
     storage.set(key, value);
     return Promise.resolve(true);
   },
 
   getItem: async (key: string): Promise<string | null> => {
+    if (isWeb && !isBrowser) return Promise.resolve(null);
+
     const value = storage.getString(key);
     return Promise.resolve(value ?? null);
   },
 
   removeItem: async (key: string): Promise<boolean> => {
+    if (isWeb && !isBrowser) return Promise.resolve(false);
+
     storage.delete(key);
     return Promise.resolve(true);
   },
@@ -26,5 +39,13 @@ const storageBackend = {
 const persistConfig = { key: "root", storage: storageBackend };
 const persistedReducer = persistReducer(persistConfig, rootReducer);
 
-export const store = configureStore({ reducer: { workouts: persistedReducer } });
+export const store = configureStore({
+  reducer: persistedReducer,
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware({
+      serializableCheck: {
+        ignoredActions: [FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER],
+      },
+    }),
+});
 export const persistor = persistStore(store);
